pyproject.toml
[project]
name = "compacted_segments"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
authors =[
    { name="Jad Abou-Chakra"}
]

dependencies = ["numpy", "torch", "warp-lang"]

[project.optional-dependencies]
test = ["pytest"]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
compacted_segments = { path = ".", editable = true }

[tool.pixi.feature.test.tasks]
test = "pytest"

[tool.pixi.environments]
test = ["test"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]